reorganize: move stuff to Swarm.Game.Cosmetic.*
