+++ src/swarm-doc/Swarm/Doc/Wiki/Cheatsheet.hs
@@ -28,8 +28,8 @@ import Swarm.Doc.Util
 import Swarm.Doc.Wiki.Matrix
 import Swarm.Doc.Wiki.Util
 import Swarm.Failure (simpleErrorHandle)
+import Swarm.Game.Cosmetic.Display (defaultChar)
 import Swarm.Game.Device qualified as D
-import Swarm.Game.Display (defaultChar)
 import Swarm.Game.Entity (Entity, EntityMap (entitiesByName), entityDisplay, entityName, loadEntities)
 import Swarm.Game.Entity qualified as E
 import Swarm.Game.Recipe (Recipe, loadRecipes, recipeCatalysts, recipeInputs, recipeOutputs, recipeTime, recipeWeight)
