+++ src/swarm-engine/Swarm/Game/Step/Const.hs
@@ -49,7 +49,7 @@ import Swarm.Effect as Effect (Time, getNow)
 import Swarm.Failure
 import Swarm.Game.Achievement.Definitions
 import Swarm.Game.CESK
-import Swarm.Game.Display
+import Swarm.Game.Cosmetic.Display
 import Swarm.Game.Entity hiding (empty, lookup, singleton, union)
 import Swarm.Game.Entity qualified as E
 import Swarm.Game.Exception
