+++ src/swarm-tui/Swarm/TUI/View/Attribute/Attr.hs
@@ -59,9 +59,9 @@ import Data.Map qualified as M
 import Data.Maybe (fromMaybe)
 import Data.Text (unpack)
 import Graphics.Vty qualified as V
-import Swarm.Game.Display (Attribute (..))
-import Swarm.Game.Entity.Cosmetic
-import Swarm.Game.Entity.Cosmetic.Assignment
+import Swarm.Game.Cosmetic.Assignment
+import Swarm.Game.Cosmetic.Color
+import Swarm.Game.Cosmetic.Display (Attribute (..))
 import Swarm.TUI.View.Attribute.Util
 
 toAttrName :: Attribute -> AttrName
